make a generic way to display dialogs opened by a button
