+++ packages/oscal-react-library/src/components/OSCALProperties.tsx
@@ -21,6 +21,7 @@ import { Property } from "@easydynamics/oscal-types";
 import { NIST_DEFAULT_NAMESPACE, namespaceOf } from "./oscal-utils/OSCALPropUtils";
 import { NotSpecifiedTypography } from "./StyledTypography";
 import { groupBy } from "../utils";
+import { ButtonLaunchedDialog } from "./ButtonLaunchedDialog";
 
 /**
  *  Helper to sort properties by their `name` field.
@@ -115,69 +116,29 @@ export const OSCALPropertiesDialog: React.FC<OSCALPropertiesDialogProps> = ({
   properties,
   title,
 }) => {
-  const [open, setOpen] = React.useState(false);
-
   const { [NIST_DEFAULT_NAMESPACE]: nist, ...thirdParties } = groupBy(properties ?? [], (prop) =>
     namespaceOf(prop.ns)
   );
-
-  const handleOpen = () => {
-    setOpen(true);
-  };
-  const handleClose = () => {
-    setOpen(false);
-  };
-
   return (
-    <>
-      <StyledTooltip title="Open Properties">
-        {
-          // This Box is necessary to ensure the tooltip is present when the button is disabled.
-          // If the Button were never disabled, the Box can be removed. This change may be made
-          // when property editing is enabled to ensure that the dialog can be opened to edit &
-          // create properties, even when none exist. In that case, even the Tooltip wrapper may
-          // not be necessary.
-        }
-        <Box display="inline">
-          <IconButton
-            color="primary"
-            size="small"
-            onClick={handleOpen}
-            aria-label="Open Properties"
-            disabled={!properties}
-          >
-            <ConstructionIcon />
-          </IconButton>
-        </Box>
-      </StyledTooltip>
-      <Dialog
-        open={open}
-        onClose={handleClose}
-        scroll="paper"
-        aria-labelledby="scroll-dialog-title"
-        aria-describedby="scroll-dialog-description"
-        maxWidth="md"
-        fullWidth
-        sx={{ maxHeight: "75em" }}
-      >
-        <DialogContent sx={{ maxHeight: "75vh" }} dividers>
-          <DialogTitle id="scroll-dialog-title">{title} Properties</DialogTitle>
-          {/* Handle NIST properties */}
-          <OSCALProperties
-            properties={nist?.sort(byName)}
-            namespace={NIST_DEFAULT_NAMESPACE}
-            key={NIST_DEFAULT_NAMESPACE}
-          />
-          {
-            /* Handle 3rd party properties */
-            Object.entries(thirdParties)
-              .sort(([key1], [key2]) => key1.localeCompare(key2))
-              .map(([key, props]) => (
-                <OSCALProperties properties={props.sort(byName)} namespace={key} key={key} />
-              ))
-          }
-        </DialogContent>
-      </Dialog>
-    </>
+    <ButtonLaunchedDialog
+      Icon={ConstructionIcon}
+      disabled={!properties}
+      title={`${title} Properties`}
+    >
+      {/* Handle NIST properties */}
+      <OSCALProperties
+        properties={nist?.sort(byName)}
+        namespace={NIST_DEFAULT_NAMESPACE}
+        key={NIST_DEFAULT_NAMESPACE}
+      />
+      {
+        /* Handle 3rd party properties */
+        Object.entries(thirdParties)
+          .sort(([key1], [key2]) => key1.localeCompare(key2))
+          .map(([key, props]) => (
+            <OSCALProperties properties={props.sort(byName)} namespace={key} key={key} />
+          ))
+      }
+    </ButtonLaunchedDialog>
   );
 };
